pyproject.toml
[project]
name = "person-lookup-agent"
version = "0.1.0"
description = "A PersonLookup agent that searches for information about individuals and creates comprehensive reports"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "certbot>=5.0.0",
    "nanda-adapter>=1.0.1",
    "openai>=1.0.0",
    "openai-agents>=0.1.0",
    "python-dotenv>=1.0.0",
    "jinja2>=3.1.0",
]
